Added LICENSE, updated the README and some documentation, added Jitpack support
