--- kbech32/common/src/kbech32/AddressFormatException.kt
@@ -1,33 +1,0 @@
-package kbech32
-
-sealed class AddressFormatException(message: String): IllegalArgumentException(message) {
-
-    /**
-     * This exception is thrown by [Base58], [Bech32] and the [PrefixedChecksummedBytes] hierarchy of
-     * classes when you try to decode data and a character isn't valid. You shouldn't allow the user to proceed in this
-     * case.
-     */
-    class InvalidCharacter(character: Char, position: Int):
-        AddressFormatException("Invalid character '$character' at position $position")
-
-    /**
-     * This exception is thrown by [Base58], [Bech32] and the [PrefixedChecksummedBytes] hierarchy of
-     * classes when you try to decode data and the data isn't of the right size. You shouldn't allow the user to proceed
-     * in this case.
-     */
-    class InvalidDataLength(message: String): AddressFormatException(message)
-
-    /**
-     * This exception is thrown by [Base58], [Bech32] and the [PrefixedChecksummedBytes] hierarchy of
-     * classes when you try to decode data and the checksum isn't valid. You shouldn't allow the user to proceed in this
-     * case.
-     */
-    class InvalidChecksum : AddressFormatException("Checksum does not validate")
-
-    /**
-     * This exception is thrown by the [PrefixedChecksummedBytes] hierarchy of classes when you try and decode an
-     * address or private key with an invalid prefix (version header or human-readable part). You shouldn't allow the
-     * user to proceed in this case.
-     */
-    class InvalidPrefix(message: String) : AddressFormatException(message)
-}
