+++ kbech32/common/src/org/komputing/kbech32/Bech32Data.kt
@@ -1,4 +1,4 @@
-package kbech32
+package org.komputing.kbech32
 
 data class Bech32Data(
     val humanReadablePart: String,
