+++ kbech32/common/test/org/komputing/kbech32/Bech32Tests.kt
@@ -1,4 +1,4 @@
-package kbech32
+package org.komputing.kbech32
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
